Multi vector colpali rag (#203)
* Added a multi-vector RAG notebook.

The notebook uses the multi-vector features added to Weaiate 1.29.0.
It builds a retrieval example based on ColPali embeddings of PDF pages
from a dataset of arXiv AI research papers. To support RAG we use the
model Qwen2.5-VL-3B-Instruct to generate answers from the top most
relevant PDF pages from our retrieval step.

* Updated multi-vector notebook with author and dependencies.

Added a multi-vectors to the table of features in the README.